Cleanup the paper adapters in the master branch
