--- worldedit-bukkit/adapters/adapter-1.19.3/build.gradle.kts
@@ -1,8 +1,0 @@
-import io.papermc.paperweight.userdev.PaperweightUserDependenciesExtension
-
-applyPaperweightAdapterConfiguration()
-
-dependencies {
-    // https://repo.papermc.io/service/rest/repository/browse/maven-public/io/papermc/paper/dev-bundle/
-    the<PaperweightUserDependenciesExtension>().paperDevBundle("1.19.3-R0.1-20221226.180038-55")
-}
